+++ CHANGELOG.md
@@ -11,6 +11,9 @@ All notable changes to this project will be documented in this file, in reverse 
   `zendframework/zend-servicemanager`.
 - [#103](https://github.com/zendframework/zend-servicemanager/pull/103) Disallowing
   test failures when running tests against PHP `7.0.*`.
+- [#103](https://github.com/zendframework/zend-servicemanager/pull/103) Improved performance
+  when dealing with registering aliases and factories via `ServiceManager#setFactory()` and
+  `ServiceManager#setAlias()`
 
 ### Deprecated